Updated SageMaker examples to take a region arg.
