--- aws/sdk/examples/sagemaker-helloworld/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "sagemaker-helloworld"
-version = "0.1.0"
-authors = ["Alistair McLean <mclean@amazon.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-env_logger = "0.8.2"
-sagemaker = {package = "aws-sdk-sagemaker", path = "../../build/aws-sdk/sagemaker"}
-tokio = { version = "1", features = ["full"] }
-
