+++ aws/sdk/examples/sagemaker/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sagemaker-code-examples"
+version = "0.1.0"
+authors = ["Alistair McLean <mclean@amazon.com>", "Doug Schwartz <dougsch@amazon.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+sagemaker = {package = "aws-sdk-sagemaker", path = "../../build/aws-sdk/sagemaker"}
+aws-types = { path = "../../build/aws-sdk/aws-types" }
+
+tokio = { version = "1", features = ["full"] }
+
+env_logger = "0.8.2"
+chrono = "0.4.19"
+structopt = { version = "0.3", default-features = false }
+tracing-subscriber = "0.2.18"
